01-color-detection/requirements.txt
opencv-python>=4.6.0
numpy>=1.24.0
Pillow>=9.2.0
